add script to populate database
